Add support to build ubuntu image for Docker

+++ packer/scripts/cleanup.sh
@@ -7,6 +7,7 @@ debian_common(){
 	apt-get -y purge $(dpkg --list | grep '^rc' | awk '{print $2}')
 	apt-get -y purge $(dpkg --list | egrep 'linux-image-[0-9]' | awk '{print $3,$2}' | sort -nr | tail -n +2 | grep -v $(uname -r) | awk '{print $2}')
 	apt-get -y clean
+	rm -rf /var/lib/apt/lists/*
 }
 
 suse_common(){
@@ -28,13 +29,17 @@ el_common(){
 }
 
 common_actions(){
+	# Clear tmp directory
+	rm -rf /tmp/*
 	# Delete unneeded files.
-	rm -f /home/vagrant/*.sh
-	# Zero out the rest of the free space using dd, then delete the written file.
-	dd if=/dev/zero of=/EMPTY bs=1M
-	rm -f /EMPTY
-	# Add `sync` so Packer doesn't quit too early, before the large file is deleted.
-	sync
+	if [ "$PACKER_PLATFROM_TYPE" = 'vagrant' ]; then
+		rm -f /home/vagrant/*.sh
+		# Zero out the rest of the free space using dd, then delete the written file.
+		dd if=/dev/zero of=/EMPTY bs=1M
+		rm -f /EMPTY
+		# Add `sync` so Packer doesn't quit too early, before the large file is deleted.
+		sync
+	fi
 	# Clear bash history
 	cat /dev/null > ~/.bash_history && history -c
 }
